move MessageDetails to chronicle-api and import
