+++ src/main/java/com/openlattice/chronicle/controllers/v2/ChronicleControllerV2.java
@@ -6,6 +6,7 @@ import com.google.common.collect.SetMultimap;
 import com.openlattice.chronicle.api.ChronicleApi;
 import com.openlattice.chronicle.data.ChronicleAppsUsageDetails;
 import com.openlattice.chronicle.data.ChronicleQuestionnaire;
+import com.openlattice.chronicle.data.MessageDetails;
 import com.openlattice.chronicle.data.ParticipationStatus;
 import com.openlattice.chronicle.services.edm.EdmCacheManager;
 import com.openlattice.chronicle.services.enrollment.EnrollmentManager;
