+++ src/main/java/com/openlattice/chronicle/services/message/MessageService.kt
@@ -4,7 +4,7 @@ import com.google.common.collect.ArrayListMultimap
 import com.google.common.collect.ImmutableSet
 import com.google.common.collect.ListMultimap
 import com.openlattice.chronicle.constants.EdmConstants
-import com.openlattice.chronicle.constants.MessageDetails
+import com.openlattice.chronicle.data.MessageDetails
 import com.openlattice.chronicle.constants.MessageOutcome
 import com.openlattice.chronicle.services.ApiCacheManager
 import com.openlattice.chronicle.services.edm.EdmCacheManager
