+++ src/main/java/com/openlattice/chronicle/services/twilio/TwilioManager.kt
@@ -1,6 +1,6 @@
 package com.openlattice.chronicle.services.twilio
 
-import com.openlattice.chronicle.constants.MessageDetails
+import com.openlattice.chronicle.data.MessageDetails
 import com.openlattice.chronicle.constants.MessageOutcome
 
 /**
