+++ src/main/java/com/openlattice/chronicle/services/twilio/TwilioService.kt
@@ -1,7 +1,7 @@
 package com.openlattice.chronicle.services.twilio
 
 import com.openlattice.chronicle.configuration.TwilioConfiguration
-import com.openlattice.chronicle.constants.MessageDetails
+import com.openlattice.chronicle.data.MessageDetails
 import com.openlattice.chronicle.constants.MessageOutcome
 import com.twilio.Twilio
 import com.twilio.exception.ApiException
